Update FlowersBirdsDrinks.html
Q1.An image titled “flowers.jpg” with a height of 150 pixels and width of 250 
pixels. If the image cannot be accessed, a message “No image available” 
should be displayed.
Q2.A hyperlink to the URL “www.mysite.com/birds.jpg”. The hyperlink 
should have the label “Click Here”.
Q3.An unordered list for listing ” Drinks “ with values Tea, Coffee, Milk.